+++ microbench/db_basic_bench.cc
@@ -9,119 +9,14 @@
 
 #include "benchmark/benchmark.h"
 #include "db/db_impl/db_impl.h"
+#include "microbench/bench_util.h"
 #include "rocksdb/db.h"
 #include "rocksdb/filter_policy.h"
 #include "rocksdb/options.h"
-#include "table/block_based/block.h"
-#include "table/block_based/block_builder.h"
 #include "util/random.h"
 #include "utilities/merge_operators.h"
 
 namespace ROCKSDB_NAMESPACE {
-
-class KeyGenerator {
- public:
-  // Generate next key
-  // buff: the caller needs to make sure there's enough space for generated key
-  // offset: to control the group of the key, 0 means normal key, 1 means
-  // non-existing key, 2 is reserved prefix_only: only return a prefix
-  Slice Next(char* buff, int8_t offset = 0, bool prefix_only = false) {
-    assert(max_key_ < std::numeric_limits<uint32_t>::max() /
-                          MULTIPLIER);  // TODO: add large key support
-
-    uint32_t k;
-    if (is_sequential_) {
-      assert(next_sequential_key_ < max_key_);
-      k = (next_sequential_key_ % max_key_) * MULTIPLIER + offset;
-      if (next_sequential_key_ + 1 == max_key_) {
-        next_sequential_key_ = 0;
-      } else {
-        next_sequential_key_++;
-      }
-    } else {
-      k = (rnd_->Next() % max_key_) * MULTIPLIER + offset;
-    }
-    // TODO: make sure the buff is large enough
-    memset(buff, 0, key_size_);
-    if (prefix_num_ > 0) {
-      uint32_t prefix = (k % prefix_num_) * MULTIPLIER + offset;
-      Encode(buff, prefix);
-      if (prefix_only) {
-        return {buff, prefix_size_};
-      }
-    }
-    Encode(buff + prefix_size_, k);
-    return {buff, key_size_};
-  }
-
-  // use internal buffer for generated key, make sure there's only one caller in
-  // single thread
-  Slice Next() { return Next(buff_); }
-
-  // user internal buffer for generated prefix
-  Slice NextPrefix() {
-    assert(prefix_num_ > 0);
-    return Next(buff_, 0, true);
-  }
-
-  // helper function to get non exist key
-  Slice NextNonExist() { return Next(buff_, 1); }
-
-  Slice MaxKey(char* buff) const {
-    memset(buff, 0xff, key_size_);
-    return {buff, key_size_};
-  }
-
-  Slice MinKey(char* buff) const {
-    memset(buff, 0, key_size_);
-    return {buff, key_size_};
-  }
-
-  // max_key: the max key that it could generate
-  // prefix_num: the max prefix number
-  // key_size: in bytes
-  explicit KeyGenerator(Random* rnd, uint64_t max_key = 100 * 1024 * 1024,
-                        size_t prefix_num = 0, size_t key_size = 10) {
-    prefix_num_ = prefix_num;
-    key_size_ = key_size;
-    max_key_ = max_key;
-    rnd_ = rnd;
-    if (prefix_num > 0) {
-      prefix_size_ = 4;  // TODO: support different prefix_size
-    }
-  }
-
-  // generate sequential keys
-  explicit KeyGenerator(uint64_t max_key = 100 * 1024 * 1024,
-                        size_t key_size = 10) {
-    key_size_ = key_size;
-    max_key_ = max_key;
-    rnd_ = nullptr;
-    is_sequential_ = true;
-  }
-
- private:
-  Random* rnd_;
-  size_t prefix_num_ = 0;
-  size_t prefix_size_ = 0;
-  size_t key_size_;
-  uint64_t max_key_;
-  bool is_sequential_ = false;
-  uint32_t next_sequential_key_ = 0;
-  char buff_[256] = {0};
-  const int MULTIPLIER = 3;
-
-  void static Encode(char* buf, uint32_t value) {
-    if (port::kLittleEndian) {
-      buf[0] = static_cast<char>((value >> 24) & 0xff);
-      buf[1] = static_cast<char>((value >> 16) & 0xff);
-      buf[2] = static_cast<char>((value >> 8) & 0xff);
-      buf[3] = static_cast<char>(value & 0xff);
-    } else {
-      memcpy(buf, &value, sizeof(value));
-    }
-  }
-};
 
 static void SetupDB(benchmark::State& state, Options& options,
                     std::unique_ptr<DB>* db,
@@ -982,85 +877,6 @@ BENCHMARK(DBGetMergeOperandsInSstFile)
     ->Threads(8)
     ->Apply(DBGetMergeOperandsInSstFileArguments);
 
-std::string GenerateKey(int primary_key, int secondary_key, int padding_size,
-                        Random* rnd) {
-  char buf[50];
-  char* p = &buf[0];
-  snprintf(buf, sizeof(buf), "%6d%4d", primary_key, secondary_key);
-  std::string k(p);
-  if (padding_size) {
-    k += rnd->RandomString(padding_size);
-  }
-
-  return k;
-}
-
-void GenerateRandomKVs(std::vector<std::string>* keys,
-                       std::vector<std::string>* values, const int from,
-                       const int len, const int step = 1,
-                       const int padding_size = 0,
-                       const int keys_share_prefix = 1) {
-  Random rnd(302);
-
-  // generate different prefix
-  for (int i = from; i < from + len; i += step) {
-    // generating keys that share the prefix
-    for (int j = 0; j < keys_share_prefix; ++j) {
-      keys->emplace_back(GenerateKey(i, j, padding_size, &rnd));
-      // 100 bytes values
-      values->emplace_back(rnd.RandomString(100));
-    }
-  }
-}
-
-// TODO: move it to different files, as it's testing an internal API
-static void DataBlockSeek(benchmark::State& state) {
-  Random rnd(301);
-  Options options = Options();
-
-  BlockBuilder builder(16, true, false,
-                       BlockBasedTableOptions::kDataBlockBinarySearch);
-
-  int num_records = 500;
-  std::vector<std::string> keys;
-  std::vector<std::string> values;
-
-  GenerateRandomKVs(&keys, &values, 0, num_records);
-
-  for (int i = 0; i < num_records; i++) {
-    std::string ukey(keys[i] + "1");
-    InternalKey ikey(ukey, 0, kTypeValue);
-    builder.Add(ikey.Encode().ToString(), values[i]);
-  }
-
-  Slice rawblock = builder.Finish();
-
-  BlockContents contents;
-  contents.data = rawblock;
-  Block reader(std::move(contents));
-
-  SetPerfLevel(kEnableTime);
-  uint64_t total = 0;
-  for (auto _ : state) {
-    DataBlockIter* iter = reader.NewDataIterator(options.comparator,
-                                                 kDisableGlobalSequenceNumber);
-    uint32_t index = rnd.Uniform(static_cast<int>(num_records));
-    std::string ukey(keys[index] + "1");
-    InternalKey ikey(ukey, 0, kTypeValue);
-    get_perf_context()->Reset();
-    bool may_exist = iter->SeekForGet(ikey.Encode().ToString());
-    if (!may_exist) {
-      state.SkipWithError("key not found");
-    }
-    total += get_perf_context()->block_seek_nanos;
-    delete iter;
-  }
-  state.counters["seek_ns"] = benchmark::Counter(
-      static_cast<double>(total), benchmark::Counter::kAvgIterations);
-}
-
-BENCHMARK(DataBlockSeek)->Iterations(1000000);
-
 static void IteratorSeek(benchmark::State& state) {
   auto compaction_style = static_cast<CompactionStyle>(state.range(0));
   uint64_t max_data = state.range(1);
@@ -1503,72 +1319,6 @@ BENCHMARK(PrefixSeek)
     ->Threads(8)
     ->Iterations(kPrefixSeekNum / 8)
     ->Apply(PrefixSeekArguments);
-
-// TODO: move it to different files, as it's testing an internal API
-static void RandomAccessFileReaderRead(benchmark::State& state) {
-  bool enable_statistics = state.range(0);
-  constexpr int kFileNum = 10;
-  auto env = Env::Default();
-  auto fs = env->GetFileSystem();
-  std::string db_path;
-  Status s = env->GetTestDirectory(&db_path);
-  if (!s.ok()) {
-    state.SkipWithError(s.ToString().c_str());
-    return;
-  }
-
-  // Setup multiple `RandomAccessFileReader`s with different parameters to be
-  // used for test
-  Random rand(301);
-  std::string fname_base =
-      db_path + kFilePathSeparator + "random-access-file-reader-read";
-  std::vector<std::unique_ptr<RandomAccessFileReader>> readers;
-  auto statistics_share = CreateDBStatistics();
-  Statistics* statistics = enable_statistics ? statistics_share.get() : nullptr;
-  for (int i = 0; i < kFileNum; i++) {
-    std::string fname = fname_base + std::to_string(i);
-    std::string content = rand.RandomString(kDefaultPageSize);
-    std::unique_ptr<WritableFile> tgt_file;
-    env->NewWritableFile(fname, &tgt_file, EnvOptions());
-    tgt_file->Append(content);
-    tgt_file->Close();
-
-    std::unique_ptr<FSRandomAccessFile> f;
-    fs->NewRandomAccessFile(fname, FileOptions(), &f, nullptr);
-    int rand_num = rand.Next() % 3;
-    auto temperature = rand_num == 0   ? Temperature::kUnknown
-                       : rand_num == 1 ? Temperature::kWarm
-                                       : Temperature::kCold;
-    readers.emplace_back(new RandomAccessFileReader(
-        std::move(f), fname, env->GetSystemClock().get(), nullptr, statistics,
-        0, nullptr, nullptr, {}, temperature, rand_num == 1));
-  }
-
-  IOOptions io_options;
-  std::unique_ptr<char[]> scratch(new char[2048]);
-  Slice result;
-  uint64_t idx = 0;
-  for (auto _ : state) {
-    s = readers[idx++ % kFileNum]->Read(io_options, 0, kDefaultPageSize / 3,
-                                        &result, scratch.get(), nullptr,
-                                        Env::IO_TOTAL);
-    if (!s.ok()) {
-      state.SkipWithError(s.ToString().c_str());
-    }
-  }
-
-  // clean up
-  for (int i = 0; i < kFileNum; i++) {
-    std::string fname = fname_base + std::to_string(i);
-    env->DeleteFile(fname);  // ignore return, okay to fail cleanup
-  }
-}
-
-BENCHMARK(RandomAccessFileReaderRead)
-    ->Iterations(1000000)
-    ->Arg(0)
-    ->Arg(1)
-    ->ArgName("enable_statistics");
 
 }  // namespace ROCKSDB_NAMESPACE
